an opportunity of more fun
Dear All,

I have uploaded right now a slightly modified version of the code distribute_v0.c, name distribute_v0_bis.c.
DO NOT WORRY about the exercise: of course it is unhaltered.

The modified version is just the opportunity of some more thoughts.
The original one, as you noticed, is not really a code that does what it claims (at least, for instance,
it does not conserve the mass - or charge, or whatever it is meant to be the quantity it spreads on the
grid, and it does not deal with the boundary conditions).

That was on purpose, just to give you a simply-enough code, and the scope of the assignment you've been
given about it is to verify whether you all are becoming familiar with some basic manipulations, basic
profiling and comparison among codes.
So, I stress it again don't be worry about the new version as for it concerns the assignment.


However, I realized that with a very small change, the willing ones among you may have a chance to have
more fun (yes, I do have a somehow strange idea of "having fun" - on such subjects, at least).

The point is the following.
The MAS (Mass ASsignment) routines that are in the code ahould actually take the "distance" argument in
units of the cell lenghth, whereas in the original code they take the distance as an absolute distance.
Because of the limits that are quoted in the routines (0.5, 1.0, 1.5), basically all the cells are
interested.

In the modified version, the distance is actually taken in units of cell length.
This ha consequences and offers the opportunity of some algorithmical optimization (the same opportunities
actually held for the original version: let's say that now they are more evident).

Just for those who wants to think about.
(In the code there are few comments marked by [change])

Cheers
Luca

p.s. Next time i'll show you a fully working code for this problem with some slightly more advanced features.